[Maps] rename connected_components/map folder to mb_map
