+++ x-pack/plugins/maps/public/classes/layers/vector_tile_layer/vector_tile_layer.js
@@ -11,7 +11,7 @@ import { isRetina } from '../../../meta';
 import {
   addSpriteSheetToMapFromImageData,
   loadSpriteSheetImageData,
-} from '../../../connected_components/map/mb/utils'; //todo move this implementation
+} from '../../../connected_components/mb_map/utils'; //todo move this implementation
 
 const MB_STYLE_TYPE_TO_OPACITY = {
   fill: ['fill-opacity'],
