+++ x-pack/plugins/maps/public/connected_components/_index.scss
@@ -2,4 +2,4 @@
 @import 'layer_panel/index';
 @import 'widget_overlay/index';
 @import 'toolbar_overlay/index';
-@import 'map/features_tooltip/index';
+@import 'mb_map/features_tooltip/index';
