+++ x-pack/plugins/maps/public/connected_components/map_container/map_container.tsx
@@ -13,7 +13,7 @@ import uuid from 'uuid/v4';
 import { Filter } from 'src/plugins/data/public';
 import { ActionExecutionContext, Action } from 'src/plugins/ui_actions/public';
 // @ts-expect-error
-import { MBMap } from '../map/mb';
+import { MBMap } from '../mb_map';
 // @ts-expect-error
 import { WidgetOverlay } from '../widget_overlay';
 // @ts-expect-error
