+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/draw_circle.ts
