+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/draw_tooltip.js
@@ -8,7 +8,7 @@ import _ from 'lodash';
 import React, { Component } from 'react';
 import { EuiPopover, EuiText } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { DRAW_TYPE } from '../../../../../common/constants';
+import { DRAW_TYPE } from '../../../../common/constants';
 
 const noop = () => {};
