+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/index.js
@@ -6,8 +6,8 @@
 
 import { connect } from 'react-redux';
 import { DrawControl } from './draw_control';
-import { updateDrawState } from '../../../../actions';
-import { getDrawState, isDrawingFilter } from '../../../../selectors/map_selectors';
+import { updateDrawState } from '../../../actions';
+import { getDrawState, isDrawingFilter } from '../../../selectors/map_selectors';
 
 function mapStateToProps(state = {}) {
   return {
