+++ x-pack/plugins/maps/public/connected_components/mb_map/features_tooltip/feature_geometry_filter_form.js
