+++ x-pack/plugins/maps/public/connected_components/mb_map/features_tooltip/tooltip_header.test.js
