+++ x-pack/plugins/maps/public/connected_components/mb_map/image_utils.js
