+++ x-pack/plugins/maps/public/connected_components/mb_map/mb.utils.test.js
@@ -5,7 +5,7 @@
  */
 
 import { removeOrphanedSourcesAndLayers } from './utils';
-import { SPATIAL_FILTERS_LAYER_ID } from '../../../../common/constants';
+import { SPATIAL_FILTERS_LAYER_ID } from '../../../common/constants';
 import _ from 'lodash';
 
 class MockMbMap {
