+++ x-pack/plugins/maps/public/connected_components/mb_map/tooltip_control/tooltip_control.test.js
